swiftgrade-cv/requirements.txt
django-dotenv==1.4.2
django-storages==1.9.1
psycopg2-binary==2.8.4
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
django-filter==2.0.0
Django==2.2
django-bulk-update==2.2.0
django-cors-headers==3.0.2
django-extensions==2.2.1
django-jet==1.0.8
rest-condition==1.0.3
ipython==7.13.0
six==1.14.0
celery==4.4.1
whitenoise==5.0.1
redis==3.3.0
drf-yasg==1.11.0
drf-friendly-errors==0.13
django-dynamic-preferences==1.7
boto3==1.12.21
gunicorn==20.0.4
Faker==4.0.2
django-admin-rangefilter==0.5.4
nameparser==1.0.6
boto==2.49.0
pycairo==1.19.1
opencv-python==3.4.4.19
fpdf==1.7.2
pypdf2==1.26.0
pdf2image==1.16.0
poppler-utils==0.1.0
requests==2.23.0
scikit-image==0.16.2
imutils==0.5.3
importlib-metadata==4.13.0
pyzbar==0.1.8
numpy==1.18.3
pandas==1.0.3
python-barcode==0.13.1
PyMuPDF==1.18.8
xxhash==2.0.0
more-itertools==8.7.0
aiohttp==3.7.4.post0